CASMINST-1074: Use the http_s3_endpoint explicitly
BSS was using the S3_ENDPOINT environment variable which will be
deprecated in the future. Change it to use the http_s3_endpoint which
is supported in the future.

Further, the S3_ENDPOINT was using https. This was causing iPXE to be
unable to fetch boot artifacts from S3. The switch to using http will
enable iPXE to fetch these boot artifacts.