Changed to search in the cached state manager data
When getting bootparameters by a name, mac, or nid it was getting all
the data in the etcd database. Reading all this data is slow when
there are a lot of bootparameters.

This commit changes it to search in the smd data which is cached in
memory. As a result it only looks up a few things in etcd.

Jira: CASMHMS-4540